Refactor: do not save as folder tree
This is the beginning of a major refactor to the script. The idea behind that change is to make the backup process as simple as possible and as error-prone as possible. The long term goal is to be able to run such script on hosted provider such as AWS Lambda, making use of the recurring scheduling features.
Instead of saving to disk by name (card name, list name or board name), the entire board is saved at once (containing all the information).